feat: hash de senha e mudanca no schema de debit

+++ finn/debit/schemas.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -16,7 +15,7 @@ class DebitSchema(OrmModel):
 
 
 class DebitCreateOrUpdateSchema(OrmModel):
-    value: Decimal
+    value: float
     category_id: int
     owner_id: int
